Move interop utils to new file.
